+++ .env.sample
@@ -1,6 +1,6 @@
 MONGO_DB_USER=
 MONGO_DB_PASSWORD=
-DATABASE_URL=
+MONGODB_DATABASE_URL=
 STATIC_TOKEN=
 
 # FRONTEND
