+++ libs/database/src/prisma/index.ts
@@ -1,2 +1,3 @@
+export * from './prisma.mongodb';
+export * from './prisma.postgresql';
 export * from './prisma.module';
-export * from './prisma.service';
