+++ libs/database/src/prisma/prisma.mongodb/index.ts
@@ -1,0 +1,1 @@
+export * from './prisma.mongodb.service';
