+++ libs/database/src/prisma/prisma.mongodb/prisma.mongodb.service.ts
@@ -1,0 +1,15 @@
+import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
+import { PrismaClient } from '@prisma/mongo/client';
+
+@Injectable()
+export class PrismaServiceMongoDB extends PrismaClient implements OnModuleInit {
+    async onModuleInit() {
+        await this.$connect();
+    }
+
+    async enableShutdownHooks(app: INestApplication) {
+        process.on('beforeExit', async () => {
+            await app.close();
+        });
+    }
+}
