+++ libs/database/src/prisma/prisma.mongodb/schema.prisma
@@ -1,11 +1,12 @@
 //This database is MONGODB area
 generator client {
   provider = "prisma-client-js"
+  output   = "../../../../../node_modules/@prisma/mongo/client"
 }
 
 datasource db {
   provider = "mongodb"
-  url      = env("DATABASE_URL")
+  url      = env("MONGODB_DATABASE_URL")
 }
 
 model Restaurants {
