+++ libs/database/src/prisma/prisma.postgresql/index.ts
@@ -1,0 +1,1 @@
+export * from './prisma.pgdb.service';
