+++ libs/database/src/restaurant/restaurant.service.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from '../prisma';
-import { Prisma } from '@prisma/client';
+import { Prisma } from '@prisma/mongo/client';
+import { PrismaServiceMongoDB } from '../prisma';
 
 @Injectable()
 export class RestaurantDBService {
-    constructor(private prisma: PrismaService) {}
+    constructor(private prisma: PrismaServiceMongoDB) {}
     async findUnique(where: Prisma.RestaurantsWhereUniqueInput) {
         const data = await this.prisma.restaurants.findUnique({
-            where,
+            where
         });
         return data;
     }
