+++ libs/database/src/users/users.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UsersDBService } from './users.service';
-import { PrismaService } from '../prisma';
+import { PrismaServiceMongoDB } from '../prisma';
 
 @Module({
-    providers: [UsersDBService, PrismaService],
+    providers: [UsersDBService, PrismaServiceMongoDB],
     exports: [UsersDBService]
 })
 export class UsersDBModule {}
